requirements.txt
Flask
tensorflow-cpu
numpy
pillow
h5py